TextToSpeech/requirements.txt
asgiref==3.5.2
comtypes==1.1.11
Django==4.0.6
djangorestframework==3.13.1
imageio==2.19.3
imutils==0.5.4
networkx==2.8.4
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
# pypiwin32==223
pytesseract==0.3.9
pyttsx3==2.90
pytz==2022.1
PyWavelets==1.3.0
# pywin32==304
scikit-image==0.19.3
scipy==1.8.1
sqlparse==0.4.2
tifffile==2022.5.4
tzdata==2022.1